+++ Assets/scripts/global/version-global.js
@@ -1,5 +1,5 @@
 export const VERSION = '0.3.0';// seismatic versioning
-export const DEV_VERSION = '14.1.0.3'; // nicodrag versioning // [total amount of updates].[Major].[minor].[changes/commits since last minor/major version change]
+export const DEV_VERSION = '14.1.0.4'; // nicodrag versioning // [total amount of updates].[Major].[minor].[changes/commits since last minor/major version change]
 
 export const DATE_VERSION = '1/??/2025';
 export const DATE_DEV_VERSION = '9/28/2024';
